Handle appendix 2 pdf
